Project Ipl/cricket-predictor-advanced/backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
xgboost
joblib
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
email-validator
requests
psycopg2-binary
python-dotenv
